# src/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Harmoniq API",
    "description": "API for managing user profiles, authors, and their publications. Auth is based on httpOnly cookies, so the Authorize button does not work here: open `POST /api/auth/login` for how to sign in from these docs, and `POST /api/auth/refresh` for what to do once your session expires.",
    "version": "1.0.0"
  },
  "servers": [
    { "url": "/" },
    { "url": "http://localhost:3000" }
  ],
...
  "paths": {
...
    "/api/articles/{id}": {
      "get": {
        "tags": ["Articles"],
        "summary": "Get an article by id",
        "parameters": [{ "$ref": "#/components/parameters/ArticleIdParam" }],
        "responses": {
          "200": {
            "description": "Article found",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "article": { "$ref": "#/components/schemas/Article" }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Article not found",
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/Error" } }
            }
          }
        }
      },
      "patch": {
        "tags": ["Articles"],
        "summary": "Update an article",
        "description": "Updates an article owned by the authenticated user. Only the owner can update it.",
        "security": [{ "sessionAuth": [], "accessTokenAuth": [] }],
        "parameters": [{ "$ref": "#/components/parameters/ArticleIdParam" }],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": { "$ref": "#/components/schemas/UpdateArticleRequest" }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Article successfully updated",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ArticleMutationResponse" }
              }
            }
          },
          "401": {
            "description": "Missing or invalid access token. Either you are not logged in, or the 15-minute access token expired: see POST /api/auth/login and POST /api/auth/refresh.",
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/Error" } }
            }
          },
          "403": {
            "description": "Access denied - not the owner of the article",
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/Error" } }
            }
          },
          "404": {
            "description": "Article not found",
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/Error" } }
            }
          }
        }
      },
      "delete": {
        "tags": ["Articles"],
        "summary": "Delete an article",
        "description": "Deletes an article owned by the authenticated user. Only the owner can delete it.",
        "security": [{ "sessionAuth": [], "accessTokenAuth": [] }],
        "parameters": [{ "$ref": "#/components/parameters/ArticleIdParam" }],
        "responses": {
          "204": {
            "description": "Article successfully deleted, no content"
          },
          "401": {
            "description": "Missing or invalid access token. Either you are not logged in, or the 15-minute access token expired: see POST /api/auth/login and POST /api/auth/refresh.",
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/Error" } }
            }
          },
          "403": {
            "description": "Not allowed to delete this article",
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/Error" } }
            }
          },
          "404": {
            "description": "Article not found",
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/Error" } }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "sessionAuth": {
        "type": "apiKey",
        "in": "cookie",
        "name": "sessionId",
        "description": "Session id cookie, set on login/register/refresh."
      },
      "accessTokenAuth": {
        "type": "apiKey",
        "in": "cookie",
        "name": "accessToken",
        "description": "Access token cookie, required for protected routes together with sessionId."
      },
      "refreshTokenAuth": {
        "type": "apiKey",
        "in": "cookie",
        "name": "refreshToken",
        "description": "Refresh token cookie, required for POST /api/auth/refresh together with sessionId."
      }
    },
    "parameters": {
...
      "ArticleIdParam": {
        "name": "id",
        "in": "path",
        "required": true,
        "schema": { "type": "string" },
        "description": "MongoDB ObjectId of the article"
      },
...
    },
    "schemas": {
      "Error": {
        "type": "object",
        "properties": {
          "message": { "type": "string" }
        }
      },
...
      "ArticleAuthor": {
        "type": "object",
        "description": "Article owner, populated by the server. Never a bare id.",
        "properties": {
          "_id": { "type": "string" },
          "name": { "type": "string" },
          "avatarUrl": { "type": "string", "nullable": true }
        }
      },
      "Article": {
        "type": "object",
        "properties": {
          "_id": { "type": "string" },
          "title": { "type": "string", "minLength": 3, "maxLength": 48 },
          "desc": { "type": "string", "maxLength": 200 },
          "article": { "type": "string", "minLength": 100, "maxLength": 4000 },
          "img": { "type": "string" },
          "rate": { "type": "number", "default": 0 },
          "date": { "type": "string", "example": "2026-08-11" },
          "category": { "type": "string", "enum": ["popular", "general"], "default": "general" },
          "ownerId": { "$ref": "#/components/schemas/ArticleAuthor" },
          "createdAt": { "type": "string", "format": "date-time" },
          "updatedAt": { "type": "string", "format": "date-time" }
        }
      },
...
      "UpdateArticleRequest": {
        "type": "object",
        "minProperties": 1,
        "properties": {
          "title": { "type": "string", "minLength": 3, "maxLength": 48 },
          "desc": { "type": "string", "maxLength": 200 },
          "article": { "type": "string", "minLength": 100, "maxLength": 4000 },
          "category": { "type": "string", "enum": ["popular", "general"] },
          "img": {
            "type": "string",
            "format": "binary",
            "description": "Article image, max 1MB"
          }
        }
      },
      "ArticleMutationResponse": {
        "type": "object",
        "properties": {
          "status": { "type": "integer", "example": 200 },
          "message": { "type": "string" },
          "data": { "$ref": "#/components/schemas/Article" }
        }
      },
...
    }
  }
}